Temp merge

+++ springapp/src/main/java/com/examly/springapp/Services/ExpenseService.java
@@ -1,11 +1,13 @@
 package com.examly.springapp.Services;
 
 import com.examly.springapp.Models.ExpenseModel;
+import com.examly.springapp.Models.UserModel;
 import com.examly.springapp.Repository.ExpenseRepository;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.List;
+import java.io.IOException;
+import org.springframework.web.multipart.MultipartFile;
 
 @Service
 public class ExpenseService {
@@ -15,30 +17,42 @@ public class ExpenseService {
     public List<ExpenseModel> findAll(){
         return this.expenseRepository.findAll();
     }
-    public ExpenseModel[] findById(int id)
+    public ExpenseModel[] findById(UserModel user)
     {
-        return this.expenseRepository.findAllExpenseByEmpId(id);
+        return this.expenseRepository.findExpenseByEmpId(user);
     }
-    public void deleteExpense(String id)
+    public String deleteExpense(String id)
     {
-        this.expenseRepository.deleteById(id);
+        if(this.expenseRepository.existsById((id)))
+        {
+            this.expenseRepository.deleteById(id);
+            return "Expense Deleted";
+        }
+        return "File doesn't exist";
+        
     }
     public String updateExpense(ExpenseModel expense)
     {
         this.expenseRepository.save(expense);
         return "Expense Updated";
     }
-    public String addExpense(ExpenseModel expense)
+    public void addExpense(ExpenseModel expense,MultipartFile file) throws IOException
     {
+        expense.setBillImage(file.getBytes());
         this.expenseRepository.save(expense);
-        return "Expense Added";
     }
-    public Long getSumOfExpenses(int id)
+    public Long getSumOfExpenses(UserModel user)
     {
-        return this.expenseRepository.findCurrentMonthExpenses(id);
+        return this.expenseRepository.findCurrentMonthExpenses(user);
     }
-    public List <ExpenseModel> getCurrentExpense(int id)
+    public List <ExpenseModel> getCurrentExpense(UserModel user)
     {
-        return this.expenseRepository.getCurrentMonthExpenses(id);
+        return this.expenseRepository.getCurrentMonthExpenses(user);
+    }
+    public ExpenseModel findByExpenseId(String expenseId) {
+        return this.expenseRepository.findExpenseByExpenseId(expenseId);
+    }
+    public ExpenseModel addExpenseTest(ExpenseModel expenseModel) {
+        return this.expenseRepository.save(expenseModel);
     }
 }
